refactored extension-loader and command-loader into combined loader.js
fixes #46
